+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
  - Interval tree from unanimity
+ - Sparse POA implementation, outsourced from unanimity
 
 ## [1.8.0] - 2020-09-23